// src/types/chart.ts
export type ChartDatasets = {
  label: string;
  data: number[];
  borderColor?: string;
  backgroundColor?: string;
};

export type Earnings = {
  january: number;
  february: number;
  march: number;
  april: number;
  may: number;
  june: number;
  july: number;
  august: number;
  september: number;
  october: number;
  november: number;
  december: number;
};